Added PropertyAndValues class. This class is used to hold the OntProperty and the values of assosiated with the TreeItem nodes.
Fixed bug that caused multiple trees to be created each time a facet was
selected.

Added Property and Value (value can be edited) in the table. Only
void:size available at the moment.

Added error dialog if the user tries to enter non-integer in the
void:size field.

Added info dialog when the user tries to select a facet without
selecting a vocabulary first.

Added file dialogs for Open... and Save As... menus (with no
functionality).

Removed some unused imports.